+++ src/util.h
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-char *create_buffer(int width, int height, int color, int modified);
+char *create_buffer(int width, int height, int color, long modified);
 char *format_timestamp(long timestamp);
 
 #endif
